chore: move import-grammars to npm run script #343
