+++ CHANGELOG.md
@@ -1,0 +1,14 @@
+# Change Log - Analysis
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](http://keepachangelog.com/)
+and this project adheres to [Semantic Versioning](http://semver.org/)
+
+## [1.0.0] - <Date>
+
+### Added
+
+-   File metric calculation: classes, functions, complexity, comments, lines_of_code, real_lines_of_code, max_nesting_level
+-   Well tested supporting languages: Go, PHP, TypeScript, TSX, Java, C#, C++, Python, JavaScript, Kotlin, JSON and YAML
+-   Not well tested languages: C, Ruby, Rust, Bash, Flow-annotated JavaScript
