+++ src/helper/helper.ts
@@ -1,8 +1,8 @@
 import fs from "node:fs/promises";
 import path from "node:path";
 import { type QueryCapture, type Tree } from "tree-sitter";
-import { type Configuration } from "../configuration.js";
-import { NodeTypeQueryStatement } from "../queries/query-statements.js";
+import { type Configuration } from "../parser/configuration.js";
+import { NodeTypeQueryStatement } from "../parser/queries/query-statements.js";
 import { type NodeTypeCategory, type NodeTypeConfig } from "./model.js";
 
 /**
