+++ src/import-grammars/import-grammar.ts
@@ -1,0 +1,3 @@
+import { updateNodeTypesMappingFile } from "./import-node-types.js";
+
+await updateNodeTypesMappingFile();
