+++ src/parser/metrics/real-lines-of-code.ts
@@ -1,8 +1,8 @@
 import { debuglog, type DebugLoggerFunction } from "node:util";
 import { type SyntaxNode, type TreeCursor } from "tree-sitter";
-import { NodeTypeCategory, type NodeTypeConfig } from "../helper/model.js";
-import { getNodeTypeNamesByCategories } from "../helper/helper.js";
-import { Language } from "../helper/language.js";
+import { NodeTypeCategory, type NodeTypeConfig } from "../../helper/model.js";
+import { getNodeTypeNamesByCategories } from "../../helper/helper.js";
+import { Language } from "../../helper/language.js";
 import { type MetricName, type Metric, type MetricResult, type ParsedFile } from "./metric.js";
 
 let dlog: DebugLoggerFunction = debuglog("metric-gardener", (logger) => {
