+++ src/parser/queries/query-builder.ts
@@ -1,6 +1,6 @@
 import { debuglog, type DebugLoggerFunction } from "node:util";
 import { Query } from "tree-sitter";
-import { type Language, languageToGrammar } from "../helper/language.js";
+import { type Language, languageToGrammar } from "../../helper/language.js";
 import { type QueryStatementInterface } from "./query-statements.js";
 
 let dlog: DebugLoggerFunction = debuglog("metric-gardener", (logger) => {
