+++ src/parser/resolver/fully-qualified-type-names/factory.ts
@@ -1,5 +1,5 @@
 import { type ParsedFile } from "../../metrics/metric.js";
-import { Language } from "../../helper/language.js";
+import { Language } from "../../../helper/language.js";
 import { type AbstractCollector } from "./abstract-collector.js";
 import { PHPCollector } from "./php-collector.js";
 import { CSharpCollector } from "./c-sharp-collector.js";
